# onboarding-control-tower/specs/001-customer-onboarding-control-tower/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Customer Onboarding Control Tower API
  version: 1.0.0
  summary: API de solo lectura del dashboard ejecutivo de onboarding (SPEC-001).
  description: |
    Contrato fuente de verdad (Art. 3 de la constitución). Todas las reglas de negocio (SLA, conversión, health,
    comparativas) se calculan en el servidor; los clientes solo representan los valores y estados recibidos.
    Los datos son sintéticos y anonimizados (REQ-001-20).
  license:
    name: Proprietary – demo
    identifier: LicenseRef-Demo
security: []
servers:
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /api/applications:
    get:
      tags: [applications]
      operationId: listApplications
      summary: Últimas solicitudes del filtro actual (más recientes primero).
      description: 'REQ-001-12, REQ-001-13; BR-004, BR-005.'
      parameters:
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/Segment'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Listado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/applications/{id}:
    get:
      tags: [applications]
      operationId: getApplication
      summary: Detalle y timeline de una solicitud.
      description: 'REQ-001-14, REQ-001-15.'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ApplicationId'
      responses:
        '200':
          description: Detalle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    Period:
      name: period
      in: query
      required: false
      description: 'Ventana móvil hasta el instante de referencia (C-001-01).'
      schema:
        $ref: '#/components/schemas/Period'
    Segment:
      name: segment
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/SegmentFilter'
  responses:
    BadRequest:
      description: Parámetros inválidos.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Recurso inexistente.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: Error inesperado (sin detalles internos).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Period:
      type: string
      enum: [today, 7d, 30d]
      default: 30d
    SegmentFilter:
      type: string
      enum: [ALL, DIGITAL, OFFICE, PARTNER]
      default: ALL
    Segment:
      type: string
      enum: [DIGITAL, OFFICE, PARTNER]
    Status:
      type: string
      enum: [STARTED, DATA_COMPLETED, VERIFYING, COMPLETED, REJECTED, EXPIRED]
    Risk:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    SlaStatus:
      type: string
      description: 'WITHIN ≤ 240 s, WARNING 241–300 s, BREACHED > 300 s; NOT_EVALUATED para rechazadas/caducadas (C-001-16).'
      enum: [WITHIN, WARNING, BREACHED, NOT_EVALUATED]
...
    Reason:
      type: string
      enum: [DOCUMENT_INVALID, IDENTITY_FAILED, OTHER, ABANDONED, INCOMPLETE]
    EventType:
      type: string
      enum: [STARTED, DATA_COMPLETED, DOCUMENT_VALIDATED, IDENTITY_VERIFIED, CUSTOMER_CREATED]
    ApplicationId:
      type: string
      pattern: '^CL-[0-9]{5}$'
      examples: [CL-10481]
...
    DateRange:
      type: object
      additionalProperties: false
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
    FilterEcho:
      type: object
      required: [period, segment, referenceNow, range, previousRange, isEmpty]
      properties:
        period:
          $ref: '#/components/schemas/Period'
        segment:
          $ref: '#/components/schemas/SegmentFilter'
        referenceNow:
          type: string
          format: date-time
          description: 'Instante de referencia de la demo (C-001-02).'
        range:
          $ref: '#/components/schemas/DateRange'
        previousRange:
          $ref: '#/components/schemas/DateRange'
        isEmpty:
          type: boolean
          description: 'true si el filtro no devuelve solicitudes en el periodo (REQ-001-19).'
...
    ApplicationSummary:
      type: object
      additionalProperties: false
      required: [id, segment, status, risk, startedAt, elapsedSec, slaStatus]
      properties:
        id:
          $ref: '#/components/schemas/ApplicationId'
        segment:
          $ref: '#/components/schemas/Segment'
        status:
          $ref: '#/components/schemas/Status'
        risk:
          $ref: '#/components/schemas/Risk'
        startedAt:
          type: string
          format: date-time
        elapsedSec:
          type: integer
          minimum: 0
        slaStatus:
          $ref: '#/components/schemas/SlaStatus'
    ApplicationList:
      allOf:
        - $ref: '#/components/schemas/FilterEcho'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              maxItems: 50
              items:
                $ref: '#/components/schemas/ApplicationSummary'
    TimelineEvent:
      type: object
      additionalProperties: false
      required: [type, occurredAt, offsetSec]
      properties:
        type:
          $ref: '#/components/schemas/EventType'
        occurredAt:
          type: string
          format: date-time
        offsetSec:
          type: integer
          minimum: 0
          description: 'Segundos desde el inicio de la solicitud.'
    ApplicationDetail:
      type: object
      additionalProperties: false
      required:
        [id, segment, status, risk, startedAt, closedAt, elapsedSec, slaStatus, rejectionReason, referenceNow, events]
      properties:
        id:
          $ref: '#/components/schemas/ApplicationId'
        segment:
          $ref: '#/components/schemas/Segment'
        status:
          $ref: '#/components/schemas/Status'
        risk:
          $ref: '#/components/schemas/Risk'
        startedAt:
          type: string
          format: date-time
        closedAt:
          type: [string, 'null']
          format: date-time
        elapsedSec:
          type: integer
          minimum: 0
        slaStatus:
          $ref: '#/components/schemas/SlaStatus'
        rejectionReason:
          oneOf:
            - $ref: '#/components/schemas/Reason'
            - type: 'null'
        referenceNow:
          type: string
          format: date-time
        events:
          type: array
          description: 'Orden cronológico ascendente (AC-001-07).'
          items:
            $ref: '#/components/schemas/TimelineEvent'
...
    Problem:
      type: object
      description: RFC 7807 Problem Details.
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string